requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-vertexai==2.0.28
google-cloud-aiplatform==1.111.0
markdown
scipy==1.15.1
scikit-learn==1.2.2
pandas==2.3.2
numpy==1.26.2
lightgbm
langchain-neo4j==0.5.0
neo4j==5.28.2
fastapi
uvicorn
python-dotenv
pydantic==2.11.7
unstructured
joblib